wip: api gateway builder using new core modeling
